Drugi commit

+++ WebVella.Erp.Plugins.Mail/MailPlugin.20190419.cs
@@ -256,7 +256,7 @@ namespace WebVella.Erp.Plugins.Mail
 				var url = "/sdk/objects/entity/l";
 				var iconClass = "fa fa-database";
 				var weight = 4;
-				var type = ((int)3);
+				var type = 3;
 				var access = new List<Guid>();
 				var entityListPages = new List<Guid>();
 				var entityCreatePages = new List<Guid>();
